MINOR: fix reading SSH output in Streams system tests (#9665)
SSH outputs in system tests originating from paramiko are bytes. However, the logger in the system tests does not accept bytes and instead throws an exception. That means, the bytes returned as SSH output from paramiko need to converted to a type that the logger (or other objects) can process.

Reviewers: Guozhang Wang <wangguoz@gmail.com>